requirements.txt
manim==0.17.3
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=9.5.0
numpy>=1.24.3
matplotlib>=3.7.1
reportlab>=3.6.12
Flask==2.0.1
Werkzeug==2.0.1
python-dateutil==2.8.2
duckduckgo-search>=3.9.3 